refactor(Login): implement MVC pattern
