+++ src/pages/login/LoginController.js
@@ -1,0 +1,5 @@
+import Login from './LoginComponent.js';
+
+export default async (target, { signIn }) => {
+  Login({ signIn }).render(target);
+};
